#8417 Added confirmation message for deleting purchase order

+++ public/admin/assets/var/demo/js/purchase-order/purchase-order-details/purchase-order-details.controller.js
@@ -2,9 +2,9 @@
 
 app.controller('PurchaseOrderDetailsController', purchaseOrderDetailsController);
 
-purchaseOrderDetailsController.$inject = ['$scope', '$state', '$attrs', '$filter', '$modal', '$timeout', '$stateParams', '$location', 'PurchaseOrderService'];
+purchaseOrderDetailsController.$inject = ['$scope', '$state', '$attrs', '$filter', '$modal', '$timeout', '$stateParams', '$location', 'SweetAlert', 'PurchaseOrderService'];
 /* @ngInject */
-function purchaseOrderDetailsController($scope, $state, $attrs, $filter, $modal, $timeout, $stateParams, $location, PurchaseOrderService) {
+function purchaseOrderDetailsController($scope, $state, $attrs, $filter, $modal, $timeout, $stateParams, $location, SweetAlert, PurchaseOrderService) {
 
     var vm = this;
 
@@ -64,12 +64,29 @@ function purchaseOrderDetailsController($scope, $state, $attrs, $filter, $modal,
 
 
     function deletePurchaseOrder(po){
-        vm.uiState.saveLoading = true;
-        PurchaseOrderService.deletePurchaseOrder(po._id).then(function(response){
-            console.log("PO deleted");
-            vm.uiState.saveLoading = false;
-            backToPurchaseOrders();
-        })
+        SweetAlert.swal({
+            title: "Are you sure?",
+            text: "You want to delete this purchase order.",
+            type: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: "Yes",
+            cancelButtonText: "Cancel",
+            closeOnConfirm: true,
+            closeOnCancel: true
+        },
+        function (isConfirm) {
+            
+            if (isConfirm) {
+                vm.uiState.saveLoading = true;
+                PurchaseOrderService.deletePurchaseOrder(po._id).then(function(response){
+                    console.log("PO deleted");
+                    vm.uiState.saveLoading = false;
+                    backToPurchaseOrders();
+                })
+            }
+        });
+        
     }
